backend/requirements.txt
Flask==2.0.1
PyMongo==3.12.0
PyJWT==2.1.0
python-dotenv==0.19.0
Flask-CORS==3.0.10